+++ src/avx2/enum_8x32.c
@@ -19,11 +19,11 @@ struct context_t {
 	const  uint32_t * const F_start;
 	__m256i * F;
 	struct solution_t buffer[512*8 + 32];
-	size_t buffer_size;
+	int64_t buffer_size;
 	uint32_t *solutions;
-	size_t n_solution_found;
-	size_t max_solutions;
-	size_t n_solutions;
+	int n_solution_found;
+	int max_solutions;
+	int n_solutions;
 	int verbose;
 };
 
@@ -62,9 +62,9 @@ static inline void STEP_2(struct context_t *context, int a, int b, uint32_t inde
    add it to the solutions. */
 static inline void FLUSH_BUFFER(struct context_t *context)
 {		
-	for (size_t i = 0; i < context->buffer_size; i++) {
+	for (int i = 0; i < context->buffer_size; i++) {
 		uint32_t x = to_gray(context->buffer[i].x);
-		for (size_t j = 0; j < 8; j++)
+		for (int j = 0; j < 8; j++)
 		if ((context->buffer[i].mask & (0x0000000f << (4 * j)))) {
 			context->solutions[context->n_solutions++] = x + j * (1 << (context->n - 3));
 			if (context->n_solutions == context->max_solutions)
@@ -75,8 +75,8 @@ static inline void FLUSH_BUFFER(struct context_t *context)
 }				
 
 // generated with L = 9
-size_t feslite_avx2_enum_8x32(size_t n, const uint32_t * const F_,
-			    uint32_t * solutions, size_t max_solutions,
+int feslite_avx2_enum_8x32(int n, const uint32_t * const F_,
+			    uint32_t * solutions, int max_solutions,
 			    bool verbose)
 {
 	struct context_t context = { .F_start = F_ };
@@ -89,11 +89,11 @@ size_t feslite_avx2_enum_8x32(size_t n, const uint32_t * const F_,
 	// RESET_COUNTER(&context);
 
 	uint64_t init_start_time = Now();
-	size_t N = idx_1(n);
+	int N = idx_1(n);
 	__m256i F[N];
 	context.F = F;
 
-	for (size_t i = 0; i < N; i++)
+	for (int i = 0; i < N; i++)
 		F[i] = _mm256_set1_epi32(F_[i]);
 
     	__m256i v0 = _mm256_set_epi32(0xffffffff, 0xffffffff, 0xffffffff, 0xffffffff, 0x00000000, 0x00000000, 0x00000000, 0x00000000);
@@ -101,16 +101,16 @@ size_t feslite_avx2_enum_8x32(size_t n, const uint32_t * const F_,
 	__m256i v2 = _mm256_set_epi32(0xffffffff, 0x00000000, 0xffffffff, 0x00000000, 0xffffffff, 0x00000000, 0xffffffff, 0x00000000);
 	 
 	F[0] ^= F[idx_1(n - 1)] & v0;
-	for (size_t i = 0; i < n - 1; i++)
+	for (int i = 0; i < n - 1; i++)
 		F[idx_1(i)] ^= F[idx_2(i, n - 1)] & v0;
 	F[0] ^= F[idx_1(n - 2)] & v1;
-	for (size_t i = 0; i < n - 2; i++)
+	for (int i = 0; i < n - 2; i++)
 		F[idx_1(i)] ^= F[idx_2(i, n - 2)] & v1;
 	F[0] ^= F[idx_1(n - 3)] & v2;
-	for (size_t i = 0; i < n - 3; i++)
+	for (int i = 0; i < n - 3; i++)
 		F[idx_1(i)] ^= F[idx_2(i, n - 3)] & v2;
 	
-	for (size_t i = 1; i < n - 3; i++)
+	for (int i = 1; i < n - 3; i++)
 		F[idx_1(i)] ^= F[idx_2(i - 1, i)];
 
 	if (verbose)
@@ -120,7 +120,7 @@ size_t feslite_avx2_enum_8x32(size_t n, const uint32_t * const F_,
 	uint64_t enumeration_start_time = Now();
 	STEP_0(&context, 0);
 
-	for (size_t idx_0 = 0; idx_0 < min(L, n - 3); idx_0++) {
+	for (int idx_0 = 0; idx_0 < min(L, n - 3); idx_0++) {
 		uint32_t w1 = (1 << idx_0);
 		STEP_1(&context, idx_1(idx_0), w1);
 		for (uint32_t i = w1 + 1; i < 2 * w1; i++) {
@@ -136,7 +136,7 @@ size_t feslite_avx2_enum_8x32(size_t n, const uint32_t * const F_,
 	}
 
 
-	for (size_t idx_0 = L; idx_0 < n - 3; idx_0++) {
+	for (int idx_0 = L; idx_0 < n - 3; idx_0++) {
 		uint32_t w1 = (1 << idx_0);
 		int alpha = idx_1(idx_0);
 		STEP_1(&context, alpha, w1);
