+++ src/generic/enum_1x32.c
@@ -16,14 +16,14 @@ struct context_t {
 	const uint32_t * const F_start;
 	uint32_t * F;
 	struct solution_t buffer[(1 << L) + 32];
-	size_t buffer_size;
+	int buffer_size;
 	uint32_t *solutions;
-	size_t max_solutions;
-	size_t n_solutions;
+	int max_solutions;
+	int n_solutions;
 
-	size_t focus[33];
-	size_t stack[32];
-	size_t sp;
+	int focus[33];
+	int stack[32];
+	int sp;
 
 	int k1;
 	int k2;
@@ -60,7 +60,7 @@ static inline void STEP_2(struct context_t *context, int a, int b, uint32_t inde
 
 static inline void FLUSH_BUFFER(struct context_t *context)
 {		
-	for (size_t i = 0; i < context->buffer_size; i++) {
+	for (int i = 0; i < context->buffer_size; i++) {
 		uint32_t x = to_gray(context->buffer[i].x);
 		context->solutions[context->n_solutions++] = x;
 		if (context->n_solutions == context->max_solutions)
@@ -81,7 +81,7 @@ static void RESET_COUNTER(struct context_t *context)
 
 static inline void UPDATE_COUNTER(struct context_t *context)
 {
-	size_t j = context->focus[0];
+	int j = context->focus[0];
 	context->focus[0] = 0;
 	context->focus[j] = context->focus[j + 1];
 	context->focus[j + 1] = j + 1;
@@ -225,8 +225,8 @@ static inline void UNROLLED_CHUNK(struct context_t *context, int alpha, uint32_t
 }
 
 // generated with L = 9
-size_t feslite_generic_enum_1x32(size_t n, const uint32_t * const F_,
-			    uint32_t * solutions, size_t max_solutions,
+int feslite_generic_enum_1x32(int n, const uint32_t * const F_,
+			    uint32_t * solutions, int max_solutions,
 			    bool verbose)
 {
 	uint64_t init_start_time = Now();
@@ -241,15 +241,15 @@ size_t feslite_generic_enum_1x32(size_t n, const uint32_t * const F_,
 
 	RESET_COUNTER(&context);
 
-	size_t N = idx_1(n);
+	int N = idx_1(n);
 	uint32_t F[N];
-	for (size_t i = 0; i < N; i++)
+	for (int i = 0; i < N; i++)
 		F[i] = F_[i];
 	context.F = F;
 
 	/* compute "derivatives" */
 	/* degree-1 terms are affected by degree-2 terms */
-	for (size_t i = 1; i < n; i++)
+	for (int i = 1; i < n; i++)
 		F[idx_1(i)] ^= F[idx_2(i - 1, i)];
 
 	if (verbose)
@@ -259,7 +259,7 @@ size_t feslite_generic_enum_1x32(size_t n, const uint32_t * const F_,
 	uint64_t enumeration_start_time = Now();
 	STEP_0(&context, 0);
 
-	for (size_t idx_0 = 0; idx_0 < min(n, L); idx_0++) {
+	for (int idx_0 = 0; idx_0 < min(n, L); idx_0++) {
 		uint32_t w1 = (1 << idx_0);
 
 		UPDATE_COUNTER(&context);
@@ -275,7 +275,7 @@ size_t feslite_generic_enum_1x32(size_t n, const uint32_t * const F_,
 
 	RESET_COUNTER(&context);
 
-	for (size_t idx_0 = L; idx_0 < n; idx_0++) {	
+	for (int idx_0 = L; idx_0 < n; idx_0++) {	
 		uint32_t w1 = (1 << idx_0);
 		
 		UPDATE_COUNTER(&context);
